Added splashart (#49)
# Description
Some last minute art for the chatroom cards.
