+++ packages/images/android-icon-background.png
