+++ packages/images/android-icon-foreground.png
