+++ packages/images/android-icon-monochrome.png
