+++ packages/images/partial-react-logo.png
