+++ packages/images/react-logo@2x.png
