+++ packages/images/splash-icon.png
